Feat : 보고서 삭제시 즐겨찾기도 같이 삭제되도록 수정
- cascade 동작 여부 재확인 필요

+++ src/resource/report/report.controller.ts
@@ -172,6 +172,6 @@ export class ReportController {
             throw new NotFoundException('해당 보고서를 찾을 수 없습니다.');
         }
 
-        return await this.reportService.deleteReport(id);
+        return await this.reportService.deleteReport({ userId, reportId: id });
     }
 }
